# Conflicts:
#	backends/gdx-backends-gwt/src/com/badlogic/gdx/backends/gwt/preloader/Preloader.java

+++ gdx/src/com/badlogic/gdx/graphics/g3d/particles/emitters/RegularEmitter.java
@@ -119,7 +119,7 @@ public class RegularEmitter extends Emitter implements Json.Serializable {
 	}
 
 	public void update () {
-		int deltaMillis = (int)(controller.deltaTime * 1000);
+		float deltaMillis = controller.deltaTime * 1000;
 
 		if (delayTimer < delay) {
 			delayTimer += deltaMillis;
